refactor: Rename ChallengeHelper to ChallengeUtils
Renamed ChallengeHelper.kt and its class to ChallengeUtils.kt to better reflect its purpose as a utility class. This change was made without moving the file to the utilities package to avoid build issues with the 'lite' flavor.
